pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "frontpipe"
version = "0.2.1"
description = "Front-of-pipe ingestion orchestrator"
authors = [{name="You"}]
dependencies = ["pyyaml", "python-dateutil"]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = ["nbformat", "pytz"]

[tool.setuptools]
packages = ["frontpipe", "frontpipe.collectors", "frontpipe.utils"]
